requirement.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
pydantic[email]
passlib[bcrypt]
python-decouple
SQLAlchemy